+++ src/Repositories/PrivateThreadRepository.php
@@ -51,7 +51,7 @@ class PrivateThreadRepository
             ->latest('updated_at')
             ->with([
                 'participants.owner',
-                'recentMessage.owner',
+                'latestMessage.owner',
                 'activeCall.participants.owner',
             ])
             ->limit($this->messenger->getThreadsIndexCount())
@@ -69,7 +69,7 @@ class PrivateThreadRepository
             ->latest('updated_at')
             ->with([
                 'participants.owner',
-                'recentMessage.owner',
+                'latestMessage.owner',
                 'activeCall.participants.owner',
             ])
             ->where('threads.updated_at', '<=', Helpers::PrecisionTime($thread->updated_at))